+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+### Changed
+- Changed `God Emperor` `Aura Of Divine Fury` from level 15 -> 14.
+- Changed `God Emperor` `Armor Saint` from level 5 -> 4.
+
 # [4.2.0] - 2023-03-20
 ### Added
 - Added `Auto Rime` OP ability (requires TabletopTweaks-Base).